Modify bloomf_test.py and merge it with bloomf.py in a Jupyter notebook
Change the test of the bloom filter in the following ways:
- Do not use given lists of words but generate random strings of various lenghts.
- Create a function to test the bloom filter. This function loops over various
sizes of string lists and bit arrays and calculates and saves the false positive
rate for each combination of string list-bit array sizes.
- Visualise the results of the test in a bar chart.

Combine the python files into one Jupyter notebook that will have to be submitted. The
Jupyter notebook also gives an initial idea of the structure of the project.