Create a new BloomFilter with two hash functions and relevant benchmark test
Original BloomFilter object contains one hash function. This does not seem
sufficient, since it has a false positive rate of over 9%, that is above
the 5% level that we require. Original BloomFilter is not modified. Rather
a new BloomFilter object is created in bloomf.py to test whether using two
hash functions improves the false positive rate. Import library hashlib.

Add a benchmark test for the new BloomFilter object. Test shows false
positive rate of 0%. However, there must be an error in the way the two
hash functions are combined in the new filter, because no string is
indicated as probably present.