Update (2nd) of alternative bloom filter, benchmarking, tests and Jupyter notebook
Re-adapt Bloom filter to initial approach of Karina that uses expected capacity and
false positive probability as inputs. This seems more in line with the standard
implementation of a Bloom filter.

Adapt benchmarking and tests accordingly.

Flesh out Jupyter notebook with a semi-final draft that includes text on theoretical
complexity.